work in progress error message system
Works but messages do not clear for each error message section in the markup.

For example, "please log in," does not disappear even after log in.
One possible fix is to have it clear programmatically, but before that, it may be worth while to use an observables to manage the messages for an immediate response instead. It will also make it so that this operation is only called in one place of the code.