updated DataManager class
Made it so that the server will send a data set under the "data" property of the json object. This eliminates the need to fetch the data again after ever server response. This also resolves a search bug where if the data had not been reloaded, searching would cause an error.