paging complete
Paging has been implemented, and works as intended. In addition, there is now a feature to toggle between the raw and styled version of the data. This will better economize the screen space.

An advanced search button has been added, but no implemented yet.  It will all a user to search the entire data set, and not just within the page.